refactor: centralize EventHandler contract
